Secure headers: add support for flat domains
